+++ RELEASES.md
@@ -3,10 +3,14 @@
 
 ## 24 / 04 / 2017 - v1.0.1
 
+Resolved issue #11
+
 - Update Save options:
   + Choose to save to a location or keep it in the temp folder
   + Use default path : workspace/images
   + Use custom path: directory / file name
+
+- Added menu for previous playblasts
 
 ## 21 / 04 / 2017 - v1.0.0